backend/quizlet/requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
install==1.3.5
Pillow==9.2.0
psycopg2==2.9.2
PyJWT==2.4.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
